% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinorBasicModel.R
\name{cosinorBasicModel}
\alias{cosinorBasicModel}
\title{cosinorBasicModel}
\usage{
cosinorBasicModel(x, window = 1, export_ts = FALSE)
}
\arguments{
\item{x}{input timeseries with missing epochs represented as "NA" values. Note
\code{length(x)*window/1440} must be an integer.}

\item{window}{epoch duration in minutes.}

\item{export_ts}{Should the original data and fitted values be exported?}
}
\value{
A list with up to two elements:
\itemize{
 \item \code{estimates}: a one-row tibble of cosinor parameters and model fit summaries.
     \itemize{
         \item The estimates of the five primary extended cosinor parameters \code{mesor_cos}, \code{amp_cos}, \code{phi_cos}.
         \item Secondary estimates from the extended cosinor model fit including the model resuduals (\code{rss_cos} and model R^2 (\code{R2_cos})).
     }
 \item \code{cosinor_ts}: a tibble containing epoch-level cosinor model fits and observed data (see \code{export_ts}).
}
}
\description{
Fit the three parameter cosinor model.
This function is a modified from the function ActExtendCosinor in ActCR.
Updates include performing the optimization on a transformed version of the
parameter space to avoid issues with boundaries.
}
\details{
The cosinor model is:
\deqn{y(t) = \mesor_{\mathrm{cos}} + \mathrm{amp}_{\mathrm{cos}}\cdot \cos(\left[ t - \phi_{\mathrm{ext}} \right]\cdot 2\pi/24) + \epsilon(t)}
}
\examples{
t <- seq(0, 1440*7,by = 1)[-1]/60
x <- 2 + 3*cos((t- 5)/24*2*pi) + rnorm(length(t),1)
cosOut <- cosinorBasicModel(x = x, window = 1,export_ts = TRUE)
cosOut$estimates
plot(
    data = cosOut$cosinor_ts,
    y~time_across_days
)
lines(
    data = cosOut$cosinor_ts,
    y_cos~time_across_days, col = "red"
)


}
