% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinorExtendedFitted.R
\name{cosinorExtendedFitted}
\alias{cosinorExtendedFitted}
\title{cosinorExtendedFitted}
\usage{
cosinorExtendedFitted(par_ext, t)
}
\arguments{
\item{par_ext}{named vector of length 5 with \code{mu_ext}, \code{gamma_ext}, \code{alpha_ext}, \code{beta_ext}, and \code{phi_ext}.}

\item{t}{time in hours since midnight.}
}
\value{
A vector of fitted values for extended cosinor model.
}
\description{
Get fitted values from [extendedCosinor::cosinorExtendedModel].
}
