% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinorExtendedModel.R
\name{cosinorExtendedModel}
\alias{cosinorExtendedModel}
\title{cosinorExtendedModel}
\usage{
cosinorExtendedModel(
  x,
  window = 1,
  export_ts = FALSE,
  export_cosinor_param = TRUE,
  export_nls_outcome = TRUE,
  maxiter = 1000,
  maxfev = 600
)
}
\arguments{
\item{x}{input timeseries with missing epochs represented as "NA" values. Note
\code{length(x)*window/1440} must be an integer.}

\item{window}{epoch duration in minutes.}

\item{export_ts}{Should the original data and fitted values be exported?}

\item{export_cosinor_param}{Should estimates from the basic three-parameter
cosinor model be added to the output?}

\item{export_nls_outcome}{Should the message and convergence code be added to the output from
minpack.lm::nls.lm.}

\item{maxiter}{Passed to [minpack.lm::nls.lm.control].}

\item{maxfev}{Passed to [minpack.lm::nls.lm.control].}
}
\value{
A list with upf to two elements:
\itemize{
 \item \code{estimates}: a one-row tibble of extended cosinor parameters, derived estimates, model fit diagnostics (see \code{export_nls_outcome}), and basic cosinor parameters/estimates (see \code{export_cosinor_param}).
     \itemize{
         \item The estimates of the five primary extended cosinor parameters \code{mu_ext}, \code{gamma_ext}, \code{alpha_ext}, \code{beta_ext}, and \code{phi_ext} (see "details").
         \item Secondary estimates from the extended cosinor model fit including the maximum, minimum, range (amplitude), midpoint (mesor), time of activity rises above and falls below the mesor (\code{up_mesor_ext} and \code{down_mesor_ext}), total daily time the curve stays above the mesor (\code{hours_above_mesor_ext}), model resuduals (\code{rss_ext}, model R^2 (\code{R2_ext}), and \code{F_pseudo_ext}).
         \item Convergence codes (\code{niter} and \code{info}) and messages (\code{message}) from minpack.lm::nls.lm.
         \item Parameters and secondary estimates from the three-parameter cosinor model: midpoint (\code{mesor_cos}), range (\code{amp_cos}), acrotime (\code{acrotime_cos}),  minimum(\code{minimum_cos}), residual sum of squares (\code{rss_cos}), and model sum of squares (\code{R2_ext}).
         \item \code{ndays}: the number of unique days of data.
     }
 \item \code{cosinor_ts}: a tibble containing epoch-level cosinor and extended cosinor model fits and observed data (see \code{export_ts}).
}
}
\description{
Fit the five parameter cosinor model from Marler et al. (2006).
This function is a modified from the function ActExtendCosinor in ActCR.
Updates include performing the optimization on a transformed version of the
parameter space to avoid issues with boundaries.
}
\details{
The model specified by Marler et al uses the following equations:
\deqn{y(t) = \mu_{\mathrm{ext}} + \gamma_{\mathrm{ext}}\cdot \mathrm{expit}\bigg(\beta_{\mathrm{ext}} \Big(\cos(\left[t - \phi_{\mathrm{ext}}\right]\cdot 2\pi/24)-\alpha_{\mathrm{ext}} \Big) \bigg) + \epsilon(t)}
}
\examples{
t <- seq(0, 1440*7,by = 1)[-1]/60
x <- sqrt(exp(2 + 3*cos((t- 5)/24*2*pi) + rnorm(length(t),1)))
cosOut <- cosinorExtendedModel(x = x, window = 1, export_ts = TRUE)
cosOut$estimates
plot(
    data = cosOut$cosinor_ts,
    y~time_across_days
)
lines(
    data = cosOut$cosinor_ts,
    y_ext~time_across_days, col = "red"
)

}
\references{
Marler MR, Gehrman P, Martin JL, Ancoli‐Israel S.
The sigmoidally transformed cosine curve: a mathematical model for circadian
rhythms with symmetric non‐sinusoidal shapes. Statistics in medicine.
2006 Nov 30;25(22):3893-904.
}
