% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinorExtendedResid.R
\name{cosinorExtendedResid}
\alias{cosinorExtendedResid}
\title{cosinorExtendedResid}
\usage{
cosinorExtendedResid(par_optim, phi_cos, Y, t)
}
\arguments{
\item{par_optim}{vector of length corresponding to transformed parameter values.
See [cosinorExtendedTransformParameters].}

\item{phi_cos}{a centering reference value for the transformation of phi.}

\item{Y}{observed data timeseries.}

\item{t}{time (in hours since midnight) corresponding to values of Y.}
}
\value{
A vector of residual values for extended cosinor model.
}
\description{
Get residual values. For use in optimization routine for model fit.
}
