% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cosinorExtendedTransformParameters.R
\name{cosinorExtendedTransformParameters}
\alias{cosinorExtendedTransformParameters}
\title{cosinorExtendedTransformParameters}
\usage{
cosinorExtendedTransformParameters(par_optim, phi_cos)
}
\arguments{
\item{par_optim}{vector of length 5 corresponding to transformed parameter values.}

\item{phi_cos}{a centering reference value for the transformation of phi.}
}
\value{
Named vector of length 5 with \code{mu_ext}, \code{gamma_ext}, \code{alpha_ext}, \code{beta_ext}, and \code{phi_ext}.
}
\description{
Transform parameters from those used in optimization routine to those specified in model (see [cosinorExtendedModel]).
}
